Improve support for multi-platform projects. Closes #641
In response to #462, a check was added to detect index stores containing multiple units for any given source file, and raise an error. The ultimate cause of the issue was the use of a custom `--index-store` path and Xcode being left open on a CI machine. Still, the check seemed reasonable to include to prevent others running into the same issue.

Some users then began running into this new error on projects that contained multi-platform targets, such as those with watchOS extensions. To workaround this, a mechanism was added that attempted to identify which index store units should be indexed by computing a target triple from the output of `xcodebuild -showBuildSettings`. Unfortunately, this approach did not work in some cases as `-showBuildSettings` does not respect the `-destination` argument, resulting in incorrect computed target triples.

There doesn't appear to be a simple way to compute a target triple ahead of time, and thus any potentially fixes will further increase the complexity of an already complex mechanism. Given that the original reason for #462 was caused by a user's particular setup, rather than an inherent issue with the index store, this change reverts all checks around the nature of units contained in the index store. I think this is a reasonable trade-off, as multi-platform projects are likely more command than the situations like the one faced in #462.

A new warning has been added when using `--index-store` to inform users that Xcode should not write to the index store while Periphery runs.